+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 
 - +[jenkinsfile], for pipeline experimentation.
 - +[express.js app], sample deployment artifact. (Apr 22 2017)
+- +[terraform, ruby, s3], beanstalk deployment definitions for node.js, still needs refinement (Apr 23 2017)